cottage-site/requirements.txt
# Core
fastapi==0.110.3
uvicorn[standard]==0.30.1
gunicorn==21.2.0
jinja2==3.1.4
python-multipart==0.0.9

# Data / ORM
SQLAlchemy==2.0.31
asyncpg==0.29.0
alembic==1.13.2

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
itsdangerous==2.2.0
email-validator==2.1.1

# Redis / Cache / PubSub
redis==5.0.4

# Settings / Validation
pydantic==2.7.4
pydantic-settings==2.3.2

# HTTP client
httpx==0.27.0

# Markdown rendering & sanitization
markdown-it-py==3.0.0
bleach==6.1.0

# Optional OAuth2 (future)
authlib==1.3.1

# Optional search client (future)
meilisearch==0.30.0

# Images (avatars)
Pillow==10.3.0
